Fix some visual studio warnings.

+++ src/common/packetsstatic_len.h
@@ -25,8 +25,8 @@
 #endif
 
 #define DEFINE_PACKET_HEADER(name, id) \
-	STATIC_ASSERT(PACKET_LEN_##id == -1 || sizeof(struct PACKET_##name) == \
-		PACKET_LEN_##id, "Wrong size PACKET_"#name); \
+	STATIC_ASSERT((int32)(PACKET_LEN_##id) == -1 || sizeof(struct PACKET_##name) == \
+		(size_t)PACKET_LEN_##id, "Wrong size PACKET_"#name); \
 	enum { HEADER_##name = id };
 
 #define packetLen(id, len) PACKET_LEN_##id = (len),
